requirements.txt
aiohttp
httpx
openai
pydub
numpy
fastembed
python-dotenv
